script/requirements.txt
tqdm
selenium
python-dotenv